requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
cytoolz==0.12.2
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
fonttools==4.41.1
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lru-dict==1.2.0
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsimonious==0.9.0
Pillow==10.0.0
protobuf==4.23.4
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunormalize==15.0.0
pywin32==306
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toolz==0.12.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
web3==6.6.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
